migrating depth-ae-convergence test
